requirements.txt
Django==6.0.1
djangorestframework==3.16.1
